bugfix: DKIM signing a particular mail
Thanks! It turns out there is a runaway regexp in Mail::DKIM
(in several of the more recent version), which your message manages
to trigger. When perl runs out of memory the process crashes.
I'll notify the author of the module.

Here is a patch to Mail::DKIM 0.33, but it should apply
to older versions too:


git-svn-id: svn://svn.code.sf.net/p/dkimproxy/code/Mail-DKIM/trunk@509 f38efd27-133c-0410-a3cc-a5f95e9cf04f