Fix regex performance issue
There is a severe regex perf hit when stripping whitespaces from email bodies when the msg has a huge # of spaces in it. We saw 10+ minutes of cpu burn for a sample real-world message. The fix is to simply swap the code that compresses whitespace to occur before stripping trailing chars. Also, it is more efficient to use my ($self, $var) = @_ instead of using shift() to peel off the $self var for any sub that is called often (i.e. once per line) and doesn’t need to pass the remaining args as array to a sub-function

Fixes RT#130710
Thanks to Todd Richmond